src/processing/audio_processor.py
"""
Audio Processor Module
Process audio for transcription
"""

import numpy as np
from typing import Optional


class AudioProcessor:
    """Audio processing utilities"""
    
    def __init__(self, target_sample_rate: int = 16000):
        self.target_sample_rate = target_sample_rate
    
    def prepare_for_whisper(self, audio_data: np.ndarray, sample_rate: int) -> np.ndarray:
        """Prepare audio for Whisper transcription"""
        # Convert to mono if stereo
        if len(audio_data.shape) > 1:
            audio_data = np.mean(audio_data, axis=1)
        
        # Resample if needed
        if sample_rate != self.target_sample_rate:
            import scipy.signal
            duration = len(audio_data) / sample_rate
            target_len = int(duration * self.target_sample_rate)
            audio_data = scipy.signal.resample(audio_data, target_len)
        
        # Normalize
        audio_data = audio_data.astype(np.float32) / np.max(np.abs(audio_data))
        
        return audio_data
    
    def extract_features(self, audio_data: np.ndarray) -> np.ndarray:
        """Extract audio features"""
        # Placeholder for feature extraction
        return audio_data
